pyproject.toml
[project]
name = "mssqlclient"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "impacket",
    "pycryptodome>=3.23.0",
]

[tool.uv.sources]
impacket = { git = "https://github.com/fortra/impacket.git" }
